[FIX]: RTA Add Iframe Sizing Issue
BCP: 2070394535

+++ src/sap.ui.rta/src/sap/ui/rta/plugin/iframe/AddIFrameDialogController.js
@@ -312,7 +312,7 @@ sap.ui.define([
 		_importIFrameSize: function(sFieldName, sSize) {
 			var aResults = sSize.split(/(px|rem|%)/);
 			if (aResults.length >= 2) {
-				this._oJSONModel.setProperty("/" + sFieldName + "/value", parseInt(aResults[0]));
+				this._oJSONModel.setProperty("/" + sFieldName + "/value", parseFloat(aResults[0]));
 				this._oJSONModel.setProperty("/" + sFieldName + "Unit/value", aResults[1]);
 			}
 		},
